Update the error handling for connection to Gitlab Server #feature #pair[hli165, ckh35]

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/service/GroupSettingsService.java
@@ -111,7 +111,7 @@ public class GroupSettingsService {
      * @param groupId current group id
      * @return true if current group setting has been saved successfully, otherwise false.
      */
-    public boolean isGroupSettingSaved(int groupSettingId, int repoId, String repoName, String repoToken, int groupId, String repoServerUrl) {
+    public boolean isGroupSettingSaved(int groupSettingId, long repoId, String repoName, String repoToken, int groupId, String repoServerUrl) {
         try {
             GroupSettings targetGroupSetting = new GroupSettings(repoId, repoName, repoToken, groupId, repoServerUrl);
             targetGroupSetting.setGroupSettingsId(groupSettingId);
